CATCH/NAME no longer multi-returns the name
Generically you can throw multi-returns, so the NAME would
need to be worked out some other way.

Overall, named throws are a bit suspicious, and probably
should be replaced with definitional throws.  If you want to
label the origin of a throw as part of the convention, that
could be part of your usage of multi-returns.