Specialize first argument of enfix drops enfix
If you say something like TYPE OF THING where OF gets its
property from the left, if you write:

    type-of: specialize :of [property: 'type]

Then TYPE-OF should not be an enfix function.  However if you
specialize out the value:

    of-thing: specialize :of [value: thing]

It should retain its enfixedness.  This does a version of that
(which doesn't take into account REORDER, but REORDER may be
getting dropped.)