Reclaim an action exec flag FULFILLING_ARG
There was an identical flag in ACTION_EXECUTOR_FLAG for fulfilling
arg to the one in the EVAL_EXECUTOR_FLAG, which was carried
through.  But it does not seem necessary, as if the action
executor turns around and calls back into the eval executor it
will always do so with the FULFILLING_ARG flag.

Removal of the flag did not seem to break any (more) tests.