Change DOING_PICKUPS from state byte to flag
An attempt was made to do ST_ACTION_DOING_PICKUPS as a state byte,
(which are not as scarce as executor flags).  But that overwrote the
case of ST_ACTION_FULFILLING_ENFIX_FROM_OUT, and sometimes the enfix
argument is actually a pickup (e.g. a refinement specialized to be
the first ordinary argument).  There's a good reason for
ENFIX_FROM_OUT to be a state byte, so this moved pickups to a flag.

(This fixes a couple of broken tests.)