Resurrect functionality of SHOVE (minus /SET)
The design of SHOVE was initially generalized across not just
pushing values into functions, but also into SET-WORD!/SET-TUPLE!.
It did so using the FEED_FLAG_NEXT_ARG_FROM_OUT and calling
into the evaluator.

At one point the sketchiness of having a feed flag talking
about taking an argument from "OUT" became apparent--since the
feed is shared across multiple frames.  It became a specific
property of an action frame, and could no longer be called from
SHOVE.  A half-baked effort to do a macro-like alternative was
made, but it wasn't a priority during the restructure and so
the tests languished.

This de-generalizes SHOVE so it only works to shove values
into functions, and so it builds an action frame directly with
no need for an evaluator frame.  As it turns out, this is simpler
and works fine for all the tests.

(The functionality needing the /SET refinement has already been
reimplemented in terms of macros for ME and MY.)