Isotopic COMMA! as distinct from Nihil
There was a concept that nihil could serve as an expression
barrier, and COMMA! would just evaluate to it.  This turned out
to be an unworkable idea... expression barriers have something
in common with empty parameter packs in terms of vaporizing in
insterstitial spots.  But they're fundamentally different when
it comes to how they're handled in some cases.

This bites the bullet and makes COMMA! evaluate to an isotopic
COMMA!, and calls that a "barrier" isotope.  Some functions
will treat these the same as NIHIL (and can refer to them as
both being ELISION?)  Other contexts treat them differently.