Bring SOURCE-ANALYSIS partially up to date
The code for running parse on C lexicals is deadly slow, and
needs to be reviewed...but not so much for speeding up PARSE3
rather to think of as a test case for optimizing UPARSE (if
it could ever hope to be useful on such a file).

So this only analyzes every 30th C file, to show that it's
working--just slow.