Standardize behavior of (read dns://)
Reading dns:// has been standardized in Ren-C to try and
get the real IP address of the system.  On Windows this is as
easy as passing null to gethostbyname().  On Linux it fabricates
a connection to Google's 8.8.8.8 DNS and asks what its side of
that connection is.

(Historically Rebol2 on Windows gave back the Windows machine
name, R3-Alpha on Linux gave an error, etc.)