Cut down cost of READABLE() in the debug build
Central to Ren-C's ability to keep moving forward with
experiments is the robust assertions and detections of when
anything is "off".  Cell format readability and writability
tests are central to that.  However, READABLE() was costing
approximately 10% of the runtime.  This trims that down to
more around 4% by noticing places where the check was
being run redundantly.