Speed up debug checks for executor flag matching
Interpretations of 8 potential flags on a frame depend on what the
executor of the frame is.  To avoid misinterpretations, the routine
for Get_Executor_Flag() took a parameter of the function pointer to
which executor you think is running when getting the flag.

As it turned out, validating the function pointer turned a very
cheap flag check into a somewhat expensive one, leading the flag
checking to consume 3% of the total runtime.  This adapts the
tactic so that optimized versions of the flag check are used in
the files of the executors themselves...so the extra checks are
only done in random other files that aren't the executor itself.