Speed up IS_LOGIC() and VAL_LOGIC()
Transitioning the datatype LOGIC! to be the isotopic states of WORD!
as ~true~ and ~false~ caused some amount of havoc in the header
files when it was done (because WORD! was defined much later than
LOGIC! initially).  So several forward definitions of functions
were needed to make it work, and it was a bit of a mess.  That mess
meant functions like IS_LOGIC() and VAL_LOGIC() weren't as optimized
as they should be due to relying on a small number of entry points.

This properly reorganizes it so that the %sys-word.h header is before
the %sys-nulled.h and %sys-logic.h headers, and does the correct
optimizations.  The result takes IS_LOGIC() from being about 1% of
the debug build runtime to closer to 0%.