Simplify/speedup cast macros
The idea of creating some safe cast() macros was a very early thing in
Ren-C.  But it accomplished this in the C++ debug build by means of
inline functions, which incurred a runtime cost (and also made
debug stepping more annoying).

This streamlines the cast macros so they don't need inline functions
most of the time.