Fix broken optimization for ACT_DETAILS
The act details array was being incorrectly assumed as always
dynamic.  This hadn't manifested as a problem because the only
time ACT_DETAILS() was being requested was when it was dynamic,
but this was somewhat coincidental.

This corrects that but works in an optimization for DETAILS_AT()
which assumes that you can only ask for a details array at a
position 1 or greater if the array is dynamic.

Also fixes a problem where a non-details array related to the
extension system was using DETAILS_AT(), and was 0-indexed instead
of 1-indexed.