Fix string performance bug w/bookmarks
The concept behind bookmarks in UTF-8 everywhere is that it has
a cache for searching in strings to help translate indices into
byte offsets--taking into account the variable size of codepoints.

An optimization trying to avoid updating the cache when requests
were made near the head of the string or tail of the string was
written incorrectly, causing the cache to not update when it
should.  This fixes the intended optimization, which had a very
significant impact on the %source-analysis.r script...making it
possible to actually run the analysis on the whole codebase (even
with a debug build) without being prohibitive.