Tweak option() to only use boolean coercibility
A point was made in the initial design of option to call out the
fact that it depended on 0 and nullptr being states that C
considered falsey--so you could get the same runtime results
whether there was a wrapper object or not.  However, just testing
against bool with assert(wrapped) vs. assert(wrapped != 0) makes
matters easier for more complex wrapped types.

Also adds a bit more explanation.