Explicitly differentiate mutable_VAL_INT64()
There was a trick where the noquote() and regular versions of
VAL_INT64() were distinguished via overload, so you could
write:

   VAL_INT64(v) = nnn;

In the C++ debug version, this would check that v was indeed
an INTEGER! and return a reference type.  So it could be used
on the left or right hand side.

This overloading distinction runs afoul of cases where you have
a smart pointer wrapper class which can coerce to either a noquote
version or a regular version.  Considering the fact that mutating
integers in this way would not even be how arbitrary precision
arithmetic would work, this sacrifices the overload so you now
have to say:

   mutable_VAL_INT64(v) = nnn;