First step at distinguishing Atom(*) type
The idea of isotopes--and in particular unstable isotopes--were
new things grafted into a codebase that was already quite shaky.
But it was deemed that there was no other way to solve some of
the language's problems, so they were pushed through in order to
further the experiment.

But unstable isotopes are particularly troublesome, and need to
be quarantined as much as possible.  This starts that process by
having a new distinguished type Atom(*) which can hold unstable
isotopes, but forces them to be decayed before copying them out.