First bootable try: ACTIONs are isotopic FRAME!s
The search for terminology for the word-reference-inert vs.
word-reference-active forms of function led to a radical idea:
what if the inert form were just FRAME!, and isotopic frames
were word-active and called "actions".

This solves the terminology problem, while reducing the total
number of types in the system.  It also grants new efficiencies,
as a specialization does not have to create a "stub" entity
to wrap the filled-in frame; the frame itself can just be an
isotope and hence be invoked directly.

There are a few speed bumps in this formulation, especially
for those used to the traditional separation of roles between
"ACTION! and FRAME!".  However there seem to be a good number
of benefits, and new users who haven't been indoctrinated in
the previous ideas would hopefully not be thrown off.