Move QUOTES_FIRST and SKIPPABLE_FIRST to varlist
These were properties that were on ACT_DETAILS(), when they
are really more properties of the interface, so should be
on paramlists and exemplars.

This is intended to help facilitate the case of when the
first argument of a function is specialized out, and there's
no new details array created--just a new exemplar.