Make enfixedness a cell flag
This moves enfixedness to using the special "type specific" bit
of frame cells.

At one time this had been done, but in such a way that the bit
was not copied as if it were a property of the action itself.
This had some trouble when doing things like MAKE FRAME! on an
action, because it could not preserve the bit.  But now that
frames and actions are unified, a frame merely needs to preserve
the bit as it would things like the binding upon specialization.

This should offer some amount of performance advantage (?)