Start generalizing details/exemplar frames
It seems the implementation of a FRAME! will either have identity
based on an interface or the implementation...depending on which
needed to be modified last.  (e.g. a SPECIALIZE changes the
interface but not the implementation, so it will give back a
FRAME! that is based on an exemplar context...while an ADAPT
changes the implementation but not the interface, so it will give
back a FRAME! that is based on a details pointer).

Given this generality, what has historically been an Action(*)
extracted from a frame will have to become something more
general from which an Exemplar() and Details() can be found.
This makes some steps toward that generalization.