Eliminate IS_ACTION() and run isotopic frames
This pushes through the changes which make VAL_ACTION() return
an Action(*) that is actually an umbrella for either type of
frame definition, e.g. either a Phase(*) or Exemplar(*).  It can
actually execute an isotopic frame made with MAKE FRAME! as a
partial specialization dispatched through a WORD!...although
there are a lot of caveats.

To make this more coherent, the current Frame(*) should be
renamed as Level(*), and Action(*) should be named Frame(*).