Rename Frame(*) to Level(*)
This is a massive change which is long overdue, that separates
the word "frame" as it pertains to the usermode FRAME! type
vs. the internal concept of Trampoline stack levels.

Additionally, it's tried to use longer and more descriptive
function names in mixed case, vs. the all caps macros.

One quirk of this commit is that it changes the lowercase `f`
used for frames to an uppercase `L` for level, simply because
the lowercase l looks too much like a 1.

Despite the large volume of changes to code and comments,
there should be no change in functionality.