Large scale type cleanup, REBSER=>Stub,Series
This commit goes through a lot of fundamental types--in particular
those starting with REB or Reb_ ... and moves them toward clearer
names that reflect the current state of the system.

Starting this push is switching over REBSER and finding places
where it wanted to refer to the stub (once called "nodes" but now
that Node term is a generalization across cells and series).  Other
places wanted to discuss the series.

The push goes across other Raw_Foo types, turning them into the
nicer looking FooT forms.  It also sorts out some build issues
in the C99 build which had slipped in recent changes.