Restore functionality of ensure() in DEBUG_CHECK_CASTS
Mistakenly, it was believed that static_cast would just do
"safe casts", e.g. not downcast.  However it does so--and
is thus less safe than dynamic_cast.

This restores the original template-based check, which costs
an inline function call in the DEBUG_CHECK_CASTS build.
Possibly it could be implemented in C++17 or C++20 without
needing to have an inline function that costs in the debug
build.