Rename TRAP_BLOCK to RESCUE_SCOPE
The naming of things in the system is such that now, TRAP is used
for catching definitional errors--cooperatively raised through the
trampoline.  So no exception handling or setjmp/longjmp is involved
with a TRAP.  Also, BLOCK can be confused with BLOCK! datatypes,
when what's really being talked about is rescuing code inside of
a C scope.

This updates the terminology, and tidies up old references to things
that no longer exist (like PUSH_TRAP())

Overall anywhere in the code that talked about "trapping errors" when
it really meant "rescuing failures" is updated.