Fix HALT invocations
The Lib(HALT) value was put into a feed for the API to execute.
Since it is a FRAME! isotope, it would turn into a quasiform of
a frame...which would evaluate to an isotope, but not run.

Passing isotopes to the API and having them produce quasiforms
is an idea that isn't very rigorous, and this shows a particular
problem with it.  It's considered useful because it can make
things like a ~null~ isotope turn around and become a null again
without having to manually meta it with rebQ().

It could be that in a similar vein, frame isotopes would splice
as plain frames... so they would execute.  But this may just be
making things worse.

In any case, this puts the Lib word HALT in instead, making the
halting work for now.