Deduplicate TG_Tick and total_eval_cycles_check
The total_eval_cycles were periodically reconciled in R3-Alpha to
avoid needing to do a countdown until signals -and- an increment
of a total count on every evaluation.  This periodic reconciliation
wasn't perfect because of problems when signals did occur, so it
was updated to be correct--and checked against a value that was
incremented in the debug build on every trampoline step.

Except...that checked value was conceptually the same thing as
the TG_tick count.  Slight adjustment made them exactly the same,
removing the need to have both.