Speed up ACT_DETAILS() as Phase_Details()
In the growing pains of switching to where FRAME! can hold an
"action-style" or "context-style" content, it introduces more
tests and dereferences on how to find certain fields.  It takes
more "hops" to get to the archetype to find the details if
you have the context style.

But if you know which kind you have, you can beeline for the
information.  As it happens, dispatchers only run when there
is an "action-style" phase in effect--because the context-style
frames have no code of their own to run (their contributions
are solely the specialized values).  So since the running code
knows it has the action style, this uses the discerned type
of Phase(*) to indicate that.

This also adds in a bit of a boost for debug builds by taking
advantage of ensure() to get const/mutable correctness for
operations like Array_Head() or Binary_At() without introducing
overloaded inline functions.  Only DEBUG_CHECK_CAST builds have
to pay, and they use a simpler identity function with one arg
and no function calls inside it.