C++ reference magic to drop mutable_FIRST_BYTE
Honoring constness made it necessary that the macros for things
like FIRST_BYTE() needed two forms... one that could handle
const input, and one that could handle mutable input and
appear on the left hand side.

    b = FIRST_BYTE(flags);  // okay if flags const
    FIRST_BYTE(flags) = b;  // should only work if flags mutable

Because this couldn't be done with plain C honoring constness, a
form `mutable_FIRST_BYTE()` was introduced for the second case.

This commit drops the idea of worrying about const correctness
in C, and only does the checks for C++...which allows the non
const variant to return a reference to a byte.

The casualty of this is that without the two forms, there isn't
one that does tests for cell READABLE() and one for WRITABLE().
Rather than wire up some convoluted test for this, we go ahead
and accept that these checks are done at a higher level than
HEART_BYTE() and QUOTE_BYTE().  A fortunate side-effect of this
is that it eliminates the need for HEART_BYTE_UNCHECKED() and
QUOTE_BYTE_UNCHECKED() since they're always unchecked.

(Generally speaking, trusting the const as the check on a cell's
readability vs. writability should take care of most of the issue
as you shouldn't be able to get mutable access on a PROTECTED
cell in the first place.)