Revive Smart Pointer Symbol(*)/String(*) behavior
The idea behind having weird Symbol(*) and Symbol(const*) all over
the place instead of just `Symbol*` and `const Symbol*` is to open
up possibilities for using smart pointers in C++ builds.

Viable classes exist using this method, that provide a large amount
of value, such as Option(...) and StackValue(...).  But ideas have
yet to fully form about what to do with Series(*), or generic Value(*)

However, if this is going to ever take shape, it helps to have a
working example--where a smart pointer exists of a size larger than
a regular pointer, and it can be used successfully.  The first test
of this was the SeriesHolder() that just did a counting of local
variables and made sure the GC didn't try to collect something
that had an outstanding reference on the stack.  This repairs that
example--which had atrophied in the face of other systemic changes.