ODBC Extension uses librebol only (no %sys-core)
The ODBC extension was very close to not using internal types and
structures...with only a few incidental things picked (like REBI64
instead of uint_64_t).

Most of the problem was being reliant on some functions for HANDLE!
which did not have API exposure.  This pushes that API through.

But another issue was that it was using the ALLOC_XXX() and FREE()
macros for getting memory out of the memory pools, as well as
using malloc().  Ideally it would be using rebAlloc() because that
gets involved with the memory bookkeeping so the system has a
more accurate accounting.  But the feature of rebUnmanage() for
memory was not implemented to give the memory blocks indefinite
lifetime beyond the frame they were declared in.

This adds the necessary features, but also takes care of one
quirk--which is that the HANDLE! cleanup functions want to be
able to free memory and extract handle properties.  This has
to be done cooperatively with the garbage collector.  So that
was implemented for these limited APIs related to freeing (any
API triggering the evaluator is not going to be possible in
the current design).