Make mod_view use libRebol, API Try=>Maybe
This takes another extension which was nearly able to be broken
off from %sys-core.h and pushes it all the way.  While attempts
were made to update the Linux GTK chooser code, it may be broken
and is not a priority to figure out how to get it built and
linked at this time.

Terminology has changed surrounding the meaning of the word
"TRY" which once had essentially the role of today's "MAYBE"
which coverts inputs to functions to a signal that indicates
the function should be tolerant of no input and return null.
TRY defuses arbitrary definitional failures, which are not
conflated with this.  Also adds a rebFreeMaybe() to allow the
callsites to more clearly indicate when they expect their
inputs to be null or not.