Replace Recycle(shutdown) mode with Sweep_Series()
In the efforts to get a "clean shutdown", Ren-C added a shutdown
flag to Recycle so it would know it was shutting down and not mark
certain recycling roots that were going away.  But in the end, what
it was really doing was freeing all the unmanaged nodes.  That
is the routine Sweep_Series(), so exporting that is easier than
trying to write a mixed-mode Recycle() that doesn't know what
things should be roots and what aren't.

Also makes mark_count a debug-only property to maintain, so marking
and unmarking are cheaper in the debug build.  Folds in a check with
the mark so that it asserts the marked/unmarked property in the
previous state.