Deduplicate NODE_FLAG_XXX in SERIES_FLAG/CELL_FLAG
At one point, what is now called a "Stub" was called a "Node", which
was confusing because the first 8 node flags were shared with cells.
This was sorted out with the creation of Stub and the idea that Node
would refer to a superclass of both Cell and Stub.  (That superclass
notion is implemented in the C++ build, and Node* is a void pointer
in C builds which don't use derived classes)

Being able to comfortably say that series and cells are nodes means
common "node functions" can be used to test for node flags, which
avoids confusion and potential mistakes of checking for flags before
the distinction between a Cell and Stub has been made (which would
violate strict aliasing).

This moves to that idea of using common routines and not duplicating
the node flags as cell and series flags...so it's clearer that the
flags are common.

Also cleans up some outdated material (like "STALE" flags which were
eliminated) and brings the remarks on rebEND up to date.  Also makes
it clearer when pools use bytes that aren't nodes by naming it
FREE_POOLUNIT_BYTE instead of FREED_SERIES_BYTE.