Scrap experimental SeriesHolder template idea
Having had significant success with smart C++ classes like Option(xxx)
and StackValue(*), the idea came up of spreading this throughout
the system.  Instead of having the source represent something that
could only be compiled as a raw pointer, using the notation DataType(*)
and DataType(const*) permitted alternate forms.

This was outlined in this article:

  http://blog.hostilefork.com/kinda-smart-pointers-in-c/

The idea of making all series subclasses use this notation was tried,
with the simplest possible application--doing reference counting of
local variables holding series.  Doing this revealed just how many
places in the system wouldn't be able to hold these smart pointers,
and how much of a burden supporting them would be.

While limited applications of the technique make sense for some
wrapper classes, it seems the series classes are too fundamentally
tied to the C implementation.  This goes ahead and removes the
speculative use of the pointer types...which also means there doesn't
have to be a second name for the raw type (e.g. ArrayT vs Array(*)).