Remove Cell(*) syntax, kill Raw_Cell/RawCell/Reb_Raw
This moves away from the idea that the basic unit of Cell would be
serviced by being able to compile in a smart pointer format, and so
loses the parentheses syntax which would be used to accomplish that.

It also gets rid of a bunch of intermediate types that were included
in order to try and work around compiler errors pertaining to memcpy()
and memset() of structures containing Cell which have the assignment
and copy constructor operators declared private in C++.  It turns out
that if you cast the target location to a void*, it subverts the
assert.  So this allowed for significant cleanup.