Revisit casts (i_cast, p_cast, constexpr...)
At one point, the complexities of cast helpers were leading to inline
functions in the debug build that were making debugging prohibitive (as
unoptimized builds do not inline things.)

It turned out there was a simple answer: constexpr.  Why I didn't think
of this before is a mystery (mostly because I didn't really know or
believe that it had been standardized in C++11, I guess).

With that, the practicality of several compile-time only constructs
means no need for certain ugly workarounds.  ensure() macro can be used
pervasively.

This also revives adherence to the safety properties of cast...based on
re-realizing how bad a generic cast can be, that converts things like
integers into pointers or vice-versa without doing so.  It adds p_cast()
for making pointers out of non-pointes, and i_cast() for making integral
types out of non-integral types.

This also does some elimination of reinterpret_cast in places where cast()
can be used in the C++ code.  Previously the thought was that ordinary
C++ would be better when C++ was available.  But this just makes the C++
look more pathological to a reader who just knows C, leading them to
not appreciate the cool type_traits stuff...it's just a distraction.