Reimagine c_cast() as "const-preserving cast"
This makes c_cast() convert from one type to another, but makes the const
of the output match the const of the input.  This helps avoid the need
to make inline functions in C++ that exist only to pick between a const
result and non-const based on the parameter.  This makes things like
Array_At() able to vanish at compile time instead of need to exist as
two different inline functions in the debug build.