Completely reedesign casts: ARR => Array, etc.
The idea behind the macros like ARR() and SER() etc. was that they'd
add extra checking in special builds, to confirm at runtime that the
bit patterns in the series were correct.

A new idea is to just make cast() itself hookable, so that this extra
code can be put in `cast(Array*, ptr)` in the special builds.  It's
longer to write out, but looks like less of a wart...and doesn't
require coming up with names for these macros for non-obvious things
(FED() for Feed?)

This also helps get away from the use of all-uppercase names for
random things in the code.

New cast mechanics are documented in %sys-casts.h.  They rely on
template partial specialization and SFINAE, and are beyond the scope
of a commit message to redescribe.