Finesse implementation of NoQuote(const Cell*)
The initial implementation of NoQuote() was as a smart pointer
container class which would accept Cells as input but not
go back without casts.  Then overloading was used to disable
operations like VAL_TYPE() that are legal on Cell but not on
NoQuote(const Cell*).

It worked.  But given that overloading was doing all the heavy
lifting, all the wrapper class was doing was implementing a
pointer that could convert in one direction only.  C++ has a more
automatic mechanism for this: the base-class and derived class
relationship.  While it may not seem completely intuitive for
NoQuote() cells to be a base class of regular cells, it works,
and doesn't have any runtime cost associated.