Take advantage of c_cast brevity to drop const in cast
The c_cast() is a const-preserving cast, which tunnels through the
constness of whatever you pass it, while performing the casts.

This makes it useful in macros that want to be polymorphic without
costing an inline function and being slow in the debug build.  But it
also allows you to write shorter code if the only reason you're
casting something const is to preserve the constness of the input.

`c_cast(Type, p)` is just shorter than `cast(const Type, p)`, and
since the constness is just sort of mandatory due to the input type,
it flows more easily in reading.