Rename REBPAR to Param, REBKEY to key
In R3-Alpha, parameter cells were called "unwords", and they shared a
layout with TYPESET! but had a symbol ID and WORD!-type of the original
parameter (WORD!, GET-WORD!, LIT-WORD!, REFINEMENT, etc.)

Full cells were used for the storage of a label list in objects as well
as functions, even though objects did not use the additional info.

Ren-C broke out the symbols into their own lists, cutting the size used
for keys in objects in 1/4.  And parameters began tracking more flags
as the behaviors became richer, to where PARAMETER! is essentially a
compressed-form of OBJECT! (similar to what EVENT! was).

In any case, Params and Keys have become pervasive, so it's time to
give them less ugly names!