Eliminate REBVAR subclass
There was an idea at one point that values representing variables
should be a subclass of REBVAL.  Reasoning was likely that there
could be some special interpretation of the cell bits that only
applied to variables--or that maybe the object a variable lived
in could be found from the value if desired.

In the years since that was introduced, no such features were
thought of.  And there are enough subclasses of Cell to create a
lot of annoyance in overloading.  This lets go of the idea.