Use macro for `inline` vs `static inline` in C++
C99 in GCC (and likely other C compilers) require you to put static on inline
functions in order to link.  C++ does not require this, and it turns out
that the savings are not insignificant..the executable 100k+ smaller
and somewhat faster when plain inline is used.  This justifies the use
of a macro, even if it seems anachronistic when `inline is part of the
C99 and C++ standards`.