Set_Series_Flag() etc. work on const series
The "const" of series is really only useful when conveying the idea
of "conceptual immutability" of the series...which is to say that the
data the series represents is const.

Beyond that, the flags and other information belong to the system, and
all the const is doing is creating a hassle when something wants to
flip the bits (e.g. the garbage collector marking a series with a bit)