Mixed-case for cell macros, e.g. IS_BLOCK() => Is_Block()
It's long been thought that the usage of all-caps in the codebase was
fairly out of hand (as if it's constantly screaming).  The lack of
rhyme or reason to why things were uppercased was also an issue.

This makes the macros for detecting cell types mixed case, which
leads to greater readability...and helps draw attention to the
macro semantics of things that are actually in all-caps (e.g.
that HEART_BYTE() can be used as the left side of an assignment)