Get rid of CELL_TO_VAL(), tidy GET_SEQUENCE_AT()
The CELL_TO_VAL() routine was a cryptic cast to legitimize taking a
NoQuote view of a cell and just considering it a regular value.  In
reality, anyone calling that operation probably wants to make a copy
and probably wants to dequotify.  So the more legitimate operation
to offer is Dequotify_Derelativize() which does both in one step,
and since it's doing both it can ignore the NoQuote() status of input.

There was also some pointless code duplication in GET_SEQUENCE_AT()
which was presumably duplicated as an optimization, but that's not
a good tradeoff.  Better to not duplicate the code.