Steps toward generalizing PAIR! to be arraylike
An unimplemented optimization is to use pairing nodes for 2-element
tuples, so it could compact things like `a.b`.  However, that
optimization has lingered as unimplemented...which means there's a
lot of fancy mechanics for pairings being maintained for the sole
purpose of implementing two-integer! pairs.

This starts the generalization process by making pairings carry an
index and a binding, to allow them to be more "arraylike".