Cleaner LINK/MISC access
The initial method for verifying LINK and MISC fields required two
different accessors...one for reading and one for writing.  The
writing one was called e.g. `mutable_LINK()`

This does some C++ magic and makes it possible for LINK() to be
either an Lvalue or Rvalue, and validate the types just the same
as before.  Getting rid of the mutable_ wart is a good thing.

There are some speedbumps, in that interaction of a template
wrapper class with other templates won't go through the normal
coercion process.  So things like ensure<> don't work with the
implementation directly, you have to extract the LINK() first
to a variable, or use a specialized node_LINK() extractor.

There may be ways to improve on this, but it's still a big step
forward in readability.