Sigmask tweaking for %mod-javascript.c
The JavaScript extension hadn't been run with full debugging on the
core before.  Once it was, problems were observed.

One problem is that the sigmask was expecting halt to be disabled in
the console, but promises weren't disabling it the way regular API
calls were.  This adds the necessary disablement to stop the assert
(though more thinking should be given to how this works.)