Don't enable NODE_FLAG_ROOT on pre-published API cells
In addition to the series stubs holding API cells carrying
NODE_FLAG_ROOT, the contained cell had NODE_FLAG_ROOT on it also
to identify it as an API cell.  This faciliates features like
if a native returns a cell with that flag set, then the action
executor knows it can release the cell.

An additional invariant that was checked was that API cells
should not contain the bit pattern representing null.  Instead,
all user-visible handles they could see which were for null
cells would specifically be conveyed by nullptr, so they could
test for null with `if` in C and not worry about freeing the
handle in null cases.

However, API cells were also used as evaluation targets before
being returned to the user by libRebol functions.  This created
a problem because during evaluation, a level's ->out pointer can
become any value--including null--making the invariant hard to
check.

This addresses the problem by making the presence of the root flag
on the cell portion of the API stub indicate that a cell is public.
The API implementations which evaluate into API cells that they
may or may not return to the user use a special initialization to
say not to put the root flag on during the evaluation targeting,
and then add the flag on before returning to the user.