Fix leaked result node in %load-r3.js
There's open questions about whether a forgiving policy on leaking
API handles in JavaScript should be implemented.  Previously the
forgiving policy was that the emscripten build wasn't made as
a debug build so you wouldn't know about leaks...

Doing a debug build found a leak in load-r3.js of the result passed
to the promise when startup was completed.