Tighten up DEBUG_UTF8_EVERYWHERE
The goal of having the C++ class to double-check the C is that it not make
the main code look too convoluted.  But there were some extra casts that
were needed due to the absence of m_cast() and c_cast() helpers on the
Utf8 wrapper.  This adds those helpers and  tidies the code somewhat.