Small amount of REBLEN=>Length/Size tinkering
The idea of using a distinct datatype to represent byte sizes vs.
element counts is a new-ish one.  Converting the code to use it is a
harrowingly process which also has decisions to make on if a different
datatype will be used for 0-based "offsets" vs. 1-based "indices".

This just does a little bit of work on the problem, until I got bored.