Restore some debug build efficiency hacks
A false belief that constexpr would compile away in debug builds
led to aggressive usage of ensure() in places that were performance
critical.  The previous methods of using differentiated names to
pick fields to avoid conflation were more practical--even though
they are less checked, just to not make debug builds suffer.

Furthering the discovery of the failings of constexpr was that the
macro `cast()` applied to compile-time constants was running a
function.  This led to nearly 20% of the debug build casting enums
and flags that should have been compile-time constants.  This
fixes that...although it may suggest that the "hooked" cast used
be named something else.