Shuffle so Any_Arraylike()/Pairlike()... in right file
There's some serious tightening up to be done on the hacked-in merge
of CHAR! and ISSUE! into a datatype which may or may not have a
dynamic allocation, but wants to act like a string type...but with
no index.  It was grafted in just well enough to work, but needs
a review of what levels of accessors do what.

This moves the code around a bit so it's at least in the right file,
and obviously wrong things like asking for the Cell_Series() of a
type that doesn't have one will draw attention.