Steps on untangling header dependencies
Core headers in the system are included in an ordering, vs. being
using include guards, to help try and expose dependencies between
layers that shouldn't exist.  There's also a lot of code in headers
in order to facilitate inlining.

This is an attempt to use the new separation of files like %sys-array.h
into %cell-array.h and %stub-array.h to organize the code more clearly.
It's able to put all the %stub-xxx.h files before all the %cell-xxx.h
files, which is a start... but there is a long way to go.