Observation about debug check casts
Having all the casts in one file helps keep them out of the main
code, but it limits what can be used to things in the %struct-xxx.h
file and exports from the .c to sys-internals.h.

This suggests a need to put the implementation code in a .cpp file
somewhere, separate from the header.  But it can wait until a
scenario motivating it arises.