Bring %cell-typeset.h up to date as %cell-parameter.h
The evolution of TYPESET! into predicate functions referenced
by array in a PARAMETER! was an organic and experimental process,
and the main issue at the time was getting the system to run at all.
Now that it does, this goes back over and updates the comments to
reflect the current state.

This updates the comments and names, bringing things up to date.

Also replaces PARAMETER_FLAG_RETURNS_VOID with the more relevant
PARAMETER_FLAG_RETURNS_NIHIL, as functions that unconditionally
return what is now called VOID are rare.