First step at parameter typecheck optimization
Implementing parameter specs as a mixture of words referring to
fundamental datatypes and functions was important.  But it slowed
down typechecking tremendously--from being just a check of a bitflag
in a 64-bit number to requiring a walk of an array and looking up
words every time.

Some improvements to speed (such as the introduction of "intrinsics")
helped matters significantly.  But it was still painfully slow, with
type checking of functions taking at least 7% of runtime.

This does a redesign of the PARAMETER! type initialization such that
it can make optimizations when it is initialized.  This just makes
it able to bypass word lookup on datatypes for up to 4 types on a
32-bit platform, and up to 8 types on a 64-bit platform.

This modest change accomplishes a 2% speedup, with hopes for more
to come.